refactor(database): mv database to mod.rs & rm database.rs;
